Day54.java

+++ Day54.java
@@ -41,7 +41,7 @@ public class Day54 {
         System.out.println("Jumlah nilai valid\t: " + jv);
         System.out.println("Jumlah nilai tidak valid: " + jtv);
 
-        double rata = tv/ jv;
+        double rata = (double) tv/ jv;
         System.out.println("Nilai valid terkecil\t: " + nk);
         System.out.println("Nilai valid terbesar\t: " + nb);
         System.out.println("Rata-rata nilai valid\t: " + rata);
